requirements.txt
astroid==2.3.1
autopep8==1.4.4
dj-database-url==0.5.0
Django==2.2.6
django-dotenv==1.4.2
django-rest-auth==0.9.5
djangorestframework==3.10.3
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
numpy==1.17.2
pandas==0.25.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pylint==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.4
wrapt==1.11.2
